+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.1.9
+- add `php-oauth-as-housekeeping` script to be run from crontab to ocassionaly
+  delete expired tokens from the database
 - no longer pass the `IniReader` object down the stack, keep all configuration 
   handling it in the `web/` scripts
 - refactor DB to accept a `PDO` object in the constructor, no longer provide